Run clippy and rustfmt on CI (#12388)
* Run clippy and rustfmt on CI

* Error on warnings and fix a couple of missed lints

* Move import inside function

* Fix unix lints

* Fix windows lints

* Missed some async tests

* Remove unneeded reference

+++ apps/desktop/desktop_native/core/src/ssh_agent/named_pipe_listener_stream.rs
@@ -61,7 +61,7 @@ impl NamedPipeServerStream {
                             }
                         };
 
-                        let peer_info = peerinfo::gather::get_peer_info(pid as u32);
+                        let peer_info = peerinfo::gather::get_peer_info(pid);
                         let peer_info = match peer_info {
                             Err(err) => {
                                 println!("Failed getting process info for pid {} {}", pid, err);
